- merge new designer story.
