+++ appglue.core/src/Form/Stories/XFormDesigner.stories.tsx
@@ -1,9 +1,15 @@
 import React from "react";
 // also exported from '@storybook/react' if you can deal with breaking changes in 6.1
 import { Story, Meta } from "@storybook/react/types-6-0";
+import styled from "styled-components";
+import { Typography } from "@material-ui/core";
+
+import {InfoIcon} from "../../CommonUI/Icon/InfoIcon";
+import {TextIcon} from "../../CommonUI/TextIcon";
 import { XFormDesigner } from "../XFormDesigner";
 import { XFormConfiguration } from "../XFormConfiguration";
 import { getFormConfig } from "../Testing/FormTestData";
+import { IAction } from "../../CommonUI/IAction";
 
 let ui = getFormConfig();
 
@@ -17,7 +23,11 @@ export default {
     component: XFormDesigner,
 } as Meta;
 
-const Template: Story<XFormDesignerProps> = (args) => <XFormDesigner {...args} />;
+const Template: Story<{
+    form: XFormConfiguration;
+    topDesignerExtensions?: IAction[],
+    bottomDesignerExtensions?: IAction[],
+}> = (args) => <XFormDesigner {...args} />;
 
 const MissingTemplate: Story<{}> = () => (
     <div>
@@ -49,3 +59,55 @@ ui = new XFormConfiguration();
 export const CloseXAction = Template.bind({}, {form: ui, onFormClose: (data: any) => {}});
 
 export const BringUpInDialogByButton = MissingTemplate.bind({}, {form: ui});
+
+
+
+const Wrapper = styled("div")<{
+    width?: number;
+    height?: number;
+    backgroundColor?: string;
+}>`
+  flex: 1;
+  display: flex;
+  flex-flow: column;
+  background: ${props => props.backgroundColor || 'transparent'}
+`;
+
+class TopExtensionPanel implements IAction {
+
+    name = 'Info';
+    icon = <InfoIcon />
+
+    renderUI() {
+        return (
+            <Wrapper backgroundColor="gray">
+                <div>
+                    <Typography>
+                        Information
+                    </Typography>
+                </div>
+            </Wrapper>
+        )
+    }
+}
+
+class BottomExtensionPanel implements IAction {
+    name = 'Doc';
+    icon = <TextIcon name="D" />
+
+    renderUI() {
+        return (
+            <Wrapper backgroundColor="green">
+                <div>
+                    <Typography>
+                        Documentation
+                    </Typography>
+                </div>
+            </Wrapper>
+        )
+    }
+}
+
+ui = getFormConfig();
+
+export const FormDesignerExtensionPanels = Template.bind({}, {form: ui, topDesignerExtensions: [new TopExtensionPanel()], bottomDesignerExtensions: [new BottomExtensionPanel()]});
